possibly fixed a compile stopper for amd64
git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@1082 70169cfe-8b10-0410-8925-dcb4b91034d8

+++ test/gwentest.c
@@ -3686,7 +3686,7 @@ int testBuffer2(int argc, char **argv) {
 
 int testFloatDouble(int argc, char **argv) {
   fprintf(stderr, "Sizeof double: %d float: %d\n",
-	  sizeof(double), sizeof(float));
+          (int)sizeof(double), (int)sizeof(float));
   return 0;
 }
